fix: Adjust size-hints of resolving entries iterator and use the upper bound in delta tree (#254)
The delta-tree is a data structure that actually heavily relies on
favorable allocation and a known amount of objects in order to
provide front and back buffers. However, this is an implementation
detail and they don't have to stay consistent at all especially
after growing the buffer by pushing to it.

Interestingly, the VecDeque internally over-allocates as well which
definitely helps the example of `as_mut_slices()`, otherwise
it could also suffer from the assertions that trigger here.