remove trait bounds to allow single-threaded applications to exist (#259)
As long as there is a feature toggle to build single-threaded
applications, they can benefit only if trait bounds don't enforce
thread Send and Sync unless truly needed.

As long as everything is goverened by the same feature toggles to
create fitting types, this will work transparently.

Previously this wasn't done as it was the assumption that code is
always capable of multi-threading even if it choses not to do it.

Now we will have types that don't actually support it due to
interior mutability, hence the relaxation of bounds where possible.