This experiment shows that working through an Arc isn't slower than through a reference (#259)
This is quite an amazing result and maybe it's just an M1 thing, but
it's clearly showing that modern CPUs can handle that pretty well.

We still keep the idea alive to make thread-safety switchable to avoid
forcing overhead on other architectures or those who don't want to use
threads.