Show that arcs can cost up to 16.5% when having 10mio ops compared to no-arc (#259)
However, that can only be caused when doing contains checks, which
probably are not usually what one does, and in standard object access
operation going through an arc (and even an arc-lock) isn't a problem.

Going through a Mutex is slow though, so it's probably better to use
read-locks when possible and upgrade them when needed.