RwLocks are preferred over Mutexes if there is a decent read-only path (#259)
For this we skip the Eager-portion and go straight to lazy loading,
while introducing a flag that should help to keep packs available
(as well as indices) indefinitely.

I imagine the coordinator of readers (upload-pack) to check for the
amount of open handles from time to time and the amount of reuse,
and replace it with a new one to clear handles.

This also means this must not be implemented as a flag per call,
but as a setting on the policy.