Implement symref (HEAD) forwarding (#543)
* Implement symref (HEAD) forwarding

Previously it was simply assumed that HEAD always points to
refs/heads/master but as it turns out that is far from true.
Now HEAD is queried per repo from the upstream and forwarded.
Unfortunately it seems that it is not avoidable to to an extra
request to the upstream server for this. (calling ls-remote)

* Silence output of merges in tests

With the new git release sometimes the merge will be done by
"ort" instead of "recursive". We don't really care about that
as the results are the same.

* Remove expect for broken refs