Don't even think about setting up test duplication for the general store (#263)
Instead this could be done by running all tests for the file store with
the type replaced to be the general store.

This means we have a normal run that runs with the default types, and
a general-store run that replaces the type used with the one for the
general store and it should all be homogenous. The latter is certainly
possible now that the loose ref store sports interior mutability
and a few platforms to get rid of some state.