repo: don't use stale view when checking for changes
I recently (0c441d9558a1) made it so we don't create an operation when
nothing changed. Soon thereafter (94e03f5ac859), I broke that when I
introduced a cache-invalidation bug when I made the filtering-out of
non-heads be lazy. This patch fixes that and also adds a test to
prevent regressions.