Find a way to handle loose dbs (#266)
This one, however, will collect them all and it's up to the handle
implementation to decide how these are to be searched.

This is only relevant for alternates, and as of now it's actually
impossible to know when the packs of one odb are done to query its loose
db. Instead, all packs of all dbs are queried first, then all loose
object stores. Maybe that's even better this way, who knows.