Sort out race condition around slots that change identity (#266)
They now carry a generation number as well, which allows 'future' slots
to be ignored by requests which are based on an old index state.

Now the potentially incorrect behaviour (getting the wrong object)
is replaced with the potential for not finding the object at all.

But maybe the latter also never happens as eventually they will queue
up around the lock to refresh path contents if they are allowed to.