Make find::Entry self-contained (#266)
That's OK because ultimately it will copy vast portions of the data.
Now we handle this by doing a mem-move inside of the vec to achieve the
same result.

However, it that's more effort than was before, for sure.
Maybe one may keep the whole entry and let the output::Entry
handle this for us, it would just have to keep track of the data offset.