fix!: do not install signal handlers by default (#336)
The previous behaviour is meant to be convenient for the casual
user even though it
ends up being surprising when used in applications that install
their own signal handlers and need more control over how the program
shuts down.

This is now fixed by **requiring an explicit `setup()`** call before
the first tempfile is created, which makes it a breaking change.