Work with std::path::* during interpolation (#331)
This means we don't necessarily enforce slashes everywhere.

Intiially the implementation was meant to be as close to `git` as
possible, even though certain requirements there stem from the
assumption that the path separtor is `/`.

In Rust, however, we now return `std::path::Path` which deals with
that for us and allows to operate on paths more comfortably and safely.

We now leverage this fully even if that means that the paths we output
aren't entirely conformant with what git would use, but it's an
implementation detail.

Lastly, the interpolated path is a transformation step so there should
be no easy way to accidentally put that version of the path back
into the file, which would be undesirable due to the various
transformations.