See if git-bitmap now compiles on 32 bit systems (#355)
Turns out that internally, it actually uses 64 bit sizes explicitly
and it was incorrect to default it to 'usize' in Rust.

Now the types are more clear.
Note that we still maintain `usize` for the callback in `for_each_bit`
for no other reason than this being what git does.